backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
paho-mqtt==2.0.0
dronekit==2.9.2
pymavlink==2.4.40
websockets==12.0
python-dotenv==1.0.0
geopy==2.4.1
aiofiles==23.2.1